+++ scilab/routines/scicos/gainblk_i32n.c
@@ -26,8 +26,8 @@ void gainblk_i32n(scicos_block *block,int flag)
      D=(double)(opar[0])*(double)(u[i]);
      t=D-(int)(D/(k))*((k));
      if ((t>=k/2)|(-(t)>=k/2))
-         {if (t>=0) (t)=(-((k/2))+abs(t-((int)((t)/((k/2))))*((k/2))));
-	  else (t)=-(-((k/2))+abs(t-((int)((t)/((k/2))))*((k/2))));}
+         {if (t>=0) (t)=(-((k/2))+abs(t-(double)((int)((t)/((k/2))))*((k/2))));
+	  else (t)=-(-((k/2))+abs(t-(double)((int)((t)/((k/2))))*((k/2))));}
      y[i]=(long)t;
     }
   }else{
@@ -42,8 +42,8 @@ void gainblk_i32n(scicos_block *block,int flag)
 		    D=D + C;}
 		    t=D-(int)(D/(k))*((k));
 		    if ((t>=k/2)|(-(t)>=k/2))
-		       {if (t>=0) (t)=(-((k/2))+abs(t-((int)((t)/((k/2))))*((k/2))));
-			else (t)=-(-((k/2))+abs(t-((int)((t)/((k/2))))*((k/2))));}
+		       {if (t>=0) (t)=(-((k/2))+abs(t-(double)((int)((t)/((k/2))))*((k/2))));
+			else (t)=-(-((k/2))+abs(t-(double)((int)((t)/((k/2))))*((k/2))));}
 		    y[jl]=(long)t;
 		  }
 	     }
